Unit Test

+++ MCTG_1/Server.cs
@@ -1,25 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Net;
-using System.Threading;
-using System.Data;
-using System.Security.Cryptography;
-using Json.Net;
+﻿using System.Net;
 using System.Net.Sockets;
-
-
+using System.Text;
+using Json.Net;
 
 namespace MCTG_1;
 
 public class Server
 {
-    private TcpListener listener;
-    private Interaction interaction;
-    private CardHandler cardHandler;
-    private Arena arena;
+    private readonly Arena arena;
+    private readonly CardHandler cardHandler;
+    private readonly Interaction interaction;
+    private readonly TcpListener listener;
+
     public Server()
     {
         listener = new TcpListener(IPAddress.Any, 10001);
@@ -34,82 +26,74 @@ public class Server
         Console.WriteLine("Server started");
         while (true)
         {
-            TcpClient client = listener.AcceptTcpClient();
-            Thread thread = new Thread(() => RequestHandler(client));
+            var client = listener.AcceptTcpClient();
+            var thread = new Thread(() => RequestHandler(client));
             thread.Start();
         }
     }
 
     private void RequestHandler(TcpClient client)
     {
-        NetworkStream stream = client.GetStream();
-        string request = "";
-        byte[] bts = new byte[2048];
+        var stream = client.GetStream();
+        var request = "";
+        var bts = new byte[2048];
         int btsR;
         do
         {
             btsR = stream.Read(bts, 0, bts.Length);
             request = Encoding.UTF8.GetString(bts, 0, btsR);
         } while (stream.DataAvailable);
-        
-        string method = request.Split(' ')[0];
-        string url = request.Split(' ')[1];
-        
+
+        var method = request.Split(' ')[0];
+        var url = request.Split(' ')[1];
+
         //split request in seperate lines
-        string[] requestLines = request.Split('\n');
+        var requestLines = request.Split('\n');
         //search for authorization header and get token
-        string token = "";
-        string userToken = "";
-        foreach (string line in requestLines)
-        {
+        var token = "";
+        var userToken = "";
+        foreach (var line in requestLines)
             if (line.Contains("Authorization"))
             {
                 token = line.Split(' ')[2];
                 userToken = token.Split("-")[0];
             }
-        }
-        
+
         //get last string in requestlines
-        string body = requestLines[requestLines.Length - 1];
-        
+        var body = requestLines[requestLines.Length - 1];
+
         Console.WriteLine(request);
         Console.WriteLine("----------------------------");
-        string answerClient = "";
-        string userAuthorization = "";
-        string contentType = "";
+        var answerClient = "";
+        var userAuthorization = "";
+        var contentType = "";
 
         if (method == "POST" && url == "/users")
         {
             // parse request body and create a new User object
-            User user = JsonNet.Deserialize<User>(body);
+            var user = JsonNet.Deserialize<User>(body);
             if (interaction.RegisterUser(user))
-            {
                 answerClient = "User " + user.Username + " registered successfully";
-            }
             else
-            {
                 answerClient = "User " + user.Username + " already exists";
-            }
-        }else if (method == "POST" && url == "/sessions")
+        }
+        else if (method == "POST" && url == "/sessions")
         {
-            string json = body;
+            var json = body;
             // parse request body and create a new User object
-            User user = JsonNet.Deserialize<User>(json);
+            var user = JsonNet.Deserialize<User>(json);
             if (interaction.Login(user))
-            {
                 answerClient = "User " + user.Username + " logged in successfully";
-            }
             else
-            {
                 answerClient = "User " + user.Username + " login failed";
-            }
-        }else if (method == "POST" && url == "/packages")
+        }
+        else if (method == "POST" && url == "/packages")
         {
             if (userToken == "admin")
             {
-                string json = body;
+                var json = body;
                 // parse request body and create a new User object
-                List<Card> package = JsonNet.Deserialize<List<Card>>(json);
+                var package = JsonNet.Deserialize<List<Card>>(json);
                 cardHandler.packages.Add(package);
                 answerClient = "Package created with " + package.Count + " cards";
                 answerClient = cardHandler.packages.Count + " packages created";
@@ -118,80 +102,80 @@ public class Server
             {
                 answerClient = "Not an admin";
             }
-
-
         }
         else if (method == "POST" && url == "/transactions/packages")
         {
-            User currentUser = new User();
+            var currentUser = new User();
             currentUser.Username = userToken;
 
             if (interaction.GetUserCoins(currentUser) >= 5)
             {
                 if (cardHandler.packages.Count > 0)
                 {
-                    
-                    cardHandler.baughtCards = (cardHandler.packages[0]);
+                    cardHandler.baughtCards = cardHandler.packages[0];
                     interaction.SaveCards(cardHandler.baughtCards, currentUser.Username);
                     cardHandler.packages.Remove(cardHandler.packages[0]);
                     interaction.UpdateCoins(currentUser);
                     answerClient += "Package bought by " + currentUser.Username;
                     answerClient += "\nPackages left: " + cardHandler.packages.Count;
                     answerClient += "\nCoins left: " + interaction.GetUserCoins(currentUser);
-                    
-                }else
+                }
+                else
                 {
                     answerClient = "No packages left";
                 }
-                
             }
             else
             {
                 answerClient = "Not enough coins";
             }
-
-        }else if (method == "GET" && url == "/cards")
+        }
+        else if (method == "GET" && url == "/cards")
         {
             if (string.IsNullOrEmpty(token))
             {
                 answerClient = "No token";
-            }else
+            }
+            else
             {
-                User currentUser = new User();
+                var currentUser = new User();
                 currentUser.Username = userToken;
                 answerClient = interaction.GetCardInfo(currentUser.Username);
             }
-        }else if (url == "/deck")
+        }
+        else if (url == "/deck")
         {
             if (method == "GET")
             {
                 if (string.IsNullOrEmpty(token))
                 {
                     answerClient = "No token";
-                }else
+                }
+                else
                 {
-                    User currentUser = new User();
+                    var currentUser = new User();
                     currentUser.Username = userToken;
                     answerClient = interaction.GetDeckInfo(currentUser.Username);
                 }
-            }else if (method == "PUT")
+            }
+            else if (method == "PUT")
             {
                 if (string.IsNullOrEmpty(token))
                 {
                     answerClient = "No token";
-                }else
+                }
+                else
                 {
-                    User currentUser = new User();
+                    var currentUser = new User();
                     currentUser.Username = userToken;
-                    List<string> deckString = JsonNet.Deserialize<List<string>>(body);
+                    var deckString = JsonNet.Deserialize<List<string>>(body);
                     if (deckString.Count != 4)
                     {
                         answerClient = "Deck must contain 4 cards";
                     }
                     else
                     {
-                        
-                        List<Card> deck = interaction.GetDeckByStrings(deckString, currentUser.Username);
+                        var deck = interaction.GetDeckByStrings(deckString, currentUser.Username);
                         if (deck.Count < 4)
                         {
                             answerClient = "Failed: One or more cards not found in your cards";
@@ -204,37 +188,41 @@ public class Server
                     }
                 }
             }
-        }else if (method == "GET" && url == "/stats")
+        }
+        else if (method == "GET" && url == "/stats")
         {
-            User currentUser = new User();
+            var currentUser = new User();
             int elo;
             currentUser.Username = userToken;
             elo = interaction.GetElo(currentUser);
             answerClient = "Elo for " + userToken + ": " + elo;
-        }else if (method == "GET" && url == "/score")
+        }
+        else if (method == "GET" && url == "/score")
         {
-            User currentUser = new User();
+            var currentUser = new User();
             currentUser.Username = userToken;
             if (interaction.UserExists(currentUser.Username))
             {
-                string allElo = "";
+                var allElo = "";
                 Console.WriteLine("User exists");
-                for (int i = 0; i < interaction.UserAmount(); i++)
+                for (var i = 0; i < interaction.UserAmount(); i++)
                 {
                     allElo += interaction.GetEloAndUsername(i);
                     allElo += "\n";
                 }
+
                 answerClient = allElo;
             }
             else
             {
                 answerClient = "User does not exist";
             }
-        }else if (url.Contains("/users/"))
+        }
+        else if (url.Contains("/users/"))
         {
             Console.WriteLine(url);
-            string[] urls = url.Split('/');
-            string dataUser = urls[2];
+            var urls = url.Split('/');
+            var dataUser = urls[2];
             Console.WriteLine(dataUser);
             if (method == "PUT")
             {
@@ -242,29 +230,32 @@ public class Server
                 {
                     //User currentuser = new User();
                     //currentuser.Username = dataUser;
-                    User user = JsonNet.Deserialize<User>(body);
+                    var user = JsonNet.Deserialize<User>(body);
                     user.Username = dataUser;
                     answerClient = "Your data is set: " + user.Name + " " + user.Bio + " " + user.Image;
                     interaction.UpdateUser(user);
-                    
-                }else
-                {
-                    answerClient = "User does not exist";
                 }
-            }else if (method == "GET")
-            {
-                if (dataUser == userToken)
-                {
-                    User currentuser = new User();
-                    currentuser.Username = dataUser;
-                    Console.WriteLine("User exists");
-                    answerClient = interaction.SelectUser(currentuser);
-                }else
+                else
                 {
                     answerClient = "User does not exist";
                 }
             }
-        }else if (url == "/tradings")
+            else if (method == "GET")
+            {
+                if (dataUser == userToken)
+                {
+                    var currentuser = new User();
+                    currentuser.Username = dataUser;
+                    Console.WriteLine("User exists");
+                    answerClient = interaction.SelectUser(currentuser);
+                }
+                else
+                {
+                    answerClient = "User does not exist";
+                }
+            }
+        }
+        else if (url == "/tradings")
         {
             if (method == "GET")
             {
@@ -272,34 +263,30 @@ public class Server
             }
             else if (method == "POST")
             {
-                string json = body;
-                Trade trade = JsonNet.Deserialize<Trade>(json);
+                var json = body;
+                var trade = JsonNet.Deserialize<Trade>(json);
                 trade.username = userToken;
 
                 answerClient = interaction.UploadTrade(trade);
             }
-            
         }
         else if (url.Contains("/tradings/"))
         {
-            string tradeId = url.Split('/')[2];
+            var tradeId = url.Split('/')[2];
 
-            if (method == "DELETE")
-            {
-                answerClient = interaction.DeleteTrade(tradeId, userToken);
-            }
+            if (method == "DELETE") answerClient = interaction.DeleteTrade(tradeId, userToken);
         }
         else if (url == "/battles")
         {
             if (method == "POST")
             {
-                User currentUser = new User();
+                var currentUser = new User();
                 currentUser.Username = userToken;
                 currentUser.Deck = interaction.GetDeckOfUser(currentUser.Username);
                 //Something wrong here!!!!!!
                 arena.AddToLobby(currentUser);
-                
-                if(arena.Lobby.Count == 2)
+
+                if (arena.Lobby.Count == 2)
                 {
                     arena.Battle();
                     answerClient = arena.PrintLog();
@@ -311,18 +298,15 @@ public class Server
             }
         }
 
-        StringBuilder response = new StringBuilder();
-        StringWriter writer = new StringWriter();
+        var response = new StringBuilder();
+        var writer = new StringWriter();
         response.Append("HTTP/1.1 200 OK\r\n");
         response.Append("Content-Type: application/json\r\n");
         response.Append("Content-Length: " + answerClient.Length + "\r\n\r\n");
         response.Append(answerClient);
-        byte[] buffer = System.Text.Encoding.UTF8.GetBytes(response.ToString());
+        var buffer = Encoding.UTF8.GetBytes(response.ToString());
         stream.Write(buffer, 0, buffer.Length);
-        
-        client.Close();
-        
-    }
-    
 
+        client.Close();
+    }
 }
